refactor(dsh-provider-usage): 阶段四——变异拓扑对齐与测试目录镜像迁移（#670）
按区块分工完成 46041b9 迁移提交的两项收尾：

D1 变异拓扑对齐（mutation-topology.json）：
- dsh-provider-usage 全部 12 段 mutate 路径从旧平铺位置更新到 §4 目标树新位置
  （apply/ -> src/apply/、errsurf -> domain2/common/、contracts+core/guards ->
  shared/contracts + domain1/pipeline/guards、entry -> apply/index + registry/
  + shared/、pipeline 段 9 文件跨 shared/registry/history/pipeline 归位、
  trend/* -> domain2/{collect,aggregate}/、report/* -> domain2/{common,schedule,
  execute}/ + apply/report-config-service、routes/* -> domain1/routes + domain2/routes）；
  段名（apply/contracts/entry/pipeline/registry/sanitize/trend-aggregate/trend-collect/
  report-schedule/report-execute/routes/errsurf）保持稳定，stryker.conf.d 文件名
  与 CI 矩阵零变化；shared/placement-math.js 为跨包共享层，不在本包迁移范围。
- testFiles 14 条登记路径同步到镜像新位置；未登记的 4 个测试文件
  （unit-detect/unit-fetch-timeout/unit-refresh-revalidate/unit-stats-service）
  维持未登记状态，不改变异口径。
- pnpm stryker:gen 重新派生全部 27 份配置、stryker:check 一致通过。

D2 测试目录镜像迁移（L1，D17 §2.1 镜像约定）：
- 18 个 unit-*.test.ts 按「测试主题与源码层对齐、文件可导航」归入
  test/unit/<层>/：apply/(apply 宿主注入)、shared/(config/contracts/v1 契约面)、
  adapters/(内置 DeepSeek 官方适配器)、client/(detect/fetch-timeout/
  refresh-revalidate，被测对象为 src/client 真源码)、common/(errsurf——
  与 src/domain2/common/ 对齐)、history/、pipeline/(signal-lock/stats-service)、
  report/(unit-report/unit-report-executor——跨 schedule/execute 主题聚簇)、
  routes/(路由层纯函数)、trend/(unit-trend 三件套聚簇，含 client/trend-math
  判定面，按任务示例归 trend/)；镜像只到层不机械复刻 domainN 前缀（topic
  聚簇更可导航）。
- 纯移动+路径同步：helpers.ts 引用 2 层、lib/index.js 与 src/* 引用 3 层、
  cross-package shared/host-utils.js 引用 5 层、join(here/pkgDir) 推导随
  深度同步；无任何测试逻辑/断言内容改动。smoke.ts 18 条 import 同步到新路径。

门禁：stryker:gen/check 全绿；pnpm --filter dsh-provider-usage test 21/21 全绿；
typecheck 当前唯一报错为并行支线 A 的 src/domain1/pipeline/interface.ts 中间态
（TS2724 StatsServiceImpl，git stash 隔离验证：无本支线改动时同样报错，与本改动
无关），待 A 收口后自然恢复。

设计决策与理由：
- 段名保持稳定：stryker.conf.d 文件名、reportName/incrementalName、CI 矩阵与
  workflow-assert 三方一致断言全部依赖段名，改名会让门禁与观察基线全量重建。
- 测试镜像映射原则：不机械照抄 src 全路径（不引入 domain1/domain2 前缀层），
  按被测源码层 + 测试主题聚簇——跨层主题（trend 数据层、report 调度/执行面）
  用主题目录，单源主题（errsurf->common、deepseek-official->adapters）对齐源码层；
  unit-v1/unit-config/unit-contract 聚在 shared/ 契约-配置面便于对照 v1/v2 契约。

关联：wingsky-1/dsh-plugin-hub#670（阶段四·目录化）
