+++ packages/dsh-provider-usage/test/unit/client/unit-fetch-timeout.test.ts
@@ -13,10 +13,10 @@ import { readFileSync, readdirSync } from "node:fs";
 import { join, dirname } from "node:path";
 import { fileURLToPath } from "node:url";
 import { execFileSync } from "node:child_process";
-import { assert } from "./helpers.ts";
+import { assert } from "../../helpers.ts";
 
 const here = dirname(fileURLToPath(import.meta.url));
-const pkgDir = join(here, "..");
+const pkgDir = join(here, "..", "..", "..");
 
 // ---------------------------------------------------------------- 源码契约：超时接线存在
