+++ packages/dsh-provider-usage/test/unit/pipeline/unit-stats-service.test.ts
@@ -13,12 +13,12 @@
 import { mkdtempSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-import { assert, pollUntil } from "./helpers.ts";
-import { normalizeConfig } from "../src/shared/config.ts";
-import { makeAdapterRegistry } from "../src/domain1/registry/registry.ts";
-import { HistoryStore } from "../src/domain1/history/history.ts";
-import { panelCacheKey } from "../src/domain1/pipeline/v2.ts";
-import { StatsService } from "../src/domain1/pipeline/stats-service.ts";
+import { assert, pollUntil } from "../../helpers.ts";
+import { normalizeConfig } from "../../../src/shared/config.ts";
+import { makeAdapterRegistry } from "../../../src/domain1/registry/registry.ts";
+import { HistoryStore } from "../../../src/domain1/history/history.ts";
+import { panelCacheKey } from "../../../src/domain1/pipeline/v2.ts";
+import { StatsService } from "../../../src/domain1/pipeline/stats-service.ts";
 
 function makeService(overrides = {}) {
   const dir = mkdtempSync(join(tmpdir(), "u-usage-stats-"));
