+++ packages/dsh-provider-usage/test/unit/report/unit-report-executor.test.ts
@@ -11,9 +11,9 @@
 import { mkdtempSync, writeFileSync, mkdirSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-import { assert } from "./helpers.ts";
-import { makeDueReportExecutor, ReportConfigService, readReportConfig, DEFAULT_REPORT_CONFIG, readLastRun } from "../lib/index.js";
-import { makeListDirs } from "../src/domain2/execute/list-dirs.ts";
+import { assert } from "../../helpers.ts";
+import { makeDueReportExecutor, ReportConfigService, readReportConfig, DEFAULT_REPORT_CONFIG, readLastRun } from "../../../lib/index.js";
+import { makeListDirs } from "../../../src/domain2/execute/list-dirs.ts";
 
 // ---- ReportConfigService：串行写链 / 内存权威 / 回调顺序 / 磁盘 roundtrip
 {
