+++ stryker.conf.d/dsh-provider-usage-contracts.json
@@ -1,8 +1,8 @@
 {
   "$schema": "../node_modules/@stryker-mutator/core/schema/stryker-schema.json",
   "mutate": [
-    "packages/dsh-provider-usage/src/contracts.ts",
-    "packages/dsh-provider-usage/src/core/guards.ts",
+    "packages/dsh-provider-usage/src/shared/contracts.ts",
+    "packages/dsh-provider-usage/src/domain1/pipeline/guards.ts",
     "!packages/dsh-provider-usage/src/client/**",
     "!packages/dsh-provider-usage/src/types.ts"
   ],
@@ -37,20 +37,20 @@
       "./scripts/test/mutation-lib-to-src-hook.mjs"
     ],
     "testFiles": [
-      "packages/dsh-provider-usage/test/unit-apply.test.ts",
-      "packages/dsh-provider-usage/test/unit-config.test.ts",
-      "packages/dsh-provider-usage/test/unit-contract.test.ts",
-      "packages/dsh-provider-usage/test/unit-deepseek-official.test.ts",
-      "packages/dsh-provider-usage/test/unit-errsurf.test.ts",
-      "packages/dsh-provider-usage/test/unit-history.test.ts",
-      "packages/dsh-provider-usage/test/unit-signal-lock.test.ts",
-      "packages/dsh-provider-usage/test/unit-v1.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend-ledger.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend-view.test.ts",
-      "packages/dsh-provider-usage/test/unit-report.test.ts",
-      "packages/dsh-provider-usage/test/unit-routes.test.ts",
-      "packages/dsh-provider-usage/test/unit-report-executor.test.ts"
+      "packages/dsh-provider-usage/test/unit/apply/unit-apply.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-config.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-contract.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-v1.test.ts",
+      "packages/dsh-provider-usage/test/unit/adapters/unit-deepseek-official.test.ts",
+      "packages/dsh-provider-usage/test/unit/history/unit-history.test.ts",
+      "packages/dsh-provider-usage/test/unit/pipeline/unit-signal-lock.test.ts",
+      "packages/dsh-provider-usage/test/unit/common/unit-errsurf.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend-ledger.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend-view.test.ts",
+      "packages/dsh-provider-usage/test/unit/report/unit-report.test.ts",
+      "packages/dsh-provider-usage/test/unit/report/unit-report-executor.test.ts",
+      "packages/dsh-provider-usage/test/unit/routes/unit-routes.test.ts"
     ]
   },
   "jsonReporter": {
@@ -58,5 +58,5 @@
   },
   "incremental": true,
   "incrementalFile": "coverage/mutation/incremental-provider-usage-contracts.json",
-  "_comment": "#342 v4：src 级文件组分段（3）——contracts+core/guards（契约面）；测试全套跑（perTest 关联）；incremental 基线随 observe 四班次重建。"
+  "_comment": "#342 v4：src 级文件组分段（3）——contracts+domain1/pipeline/guards（契约面）；测试全套跑（perTest 关联）；incremental 基线随 observe 四班次重建。"
 }
