+++ stryker.conf.d/dsh-provider-usage-entry.json
@@ -1,10 +1,10 @@
 {
   "$schema": "../node_modules/@stryker-mutator/core/schema/stryker-schema.json",
   "mutate": [
-    "packages/dsh-provider-usage/src/index.ts",
-    "packages/dsh-provider-usage/src/user-adapters.ts",
-    "packages/dsh-provider-usage/src/ui-config.ts",
-    "packages/dsh-provider-usage/src/config.ts",
+    "packages/dsh-provider-usage/src/apply/index.ts",
+    "packages/dsh-provider-usage/src/domain1/registry/user-adapters.ts",
+    "packages/dsh-provider-usage/src/shared/ui-config.ts",
+    "packages/dsh-provider-usage/src/shared/config.ts",
     "!packages/dsh-provider-usage/src/client/**",
     "!packages/dsh-provider-usage/src/types.ts"
   ],
@@ -39,20 +39,20 @@
       "./scripts/test/mutation-lib-to-src-hook.mjs"
     ],
     "testFiles": [
-      "packages/dsh-provider-usage/test/unit-apply.test.ts",
-      "packages/dsh-provider-usage/test/unit-config.test.ts",
-      "packages/dsh-provider-usage/test/unit-contract.test.ts",
-      "packages/dsh-provider-usage/test/unit-deepseek-official.test.ts",
-      "packages/dsh-provider-usage/test/unit-errsurf.test.ts",
-      "packages/dsh-provider-usage/test/unit-history.test.ts",
-      "packages/dsh-provider-usage/test/unit-signal-lock.test.ts",
-      "packages/dsh-provider-usage/test/unit-v1.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend-ledger.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend-view.test.ts",
-      "packages/dsh-provider-usage/test/unit-report.test.ts",
-      "packages/dsh-provider-usage/test/unit-routes.test.ts",
-      "packages/dsh-provider-usage/test/unit-report-executor.test.ts"
+      "packages/dsh-provider-usage/test/unit/apply/unit-apply.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-config.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-contract.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-v1.test.ts",
+      "packages/dsh-provider-usage/test/unit/adapters/unit-deepseek-official.test.ts",
+      "packages/dsh-provider-usage/test/unit/history/unit-history.test.ts",
+      "packages/dsh-provider-usage/test/unit/pipeline/unit-signal-lock.test.ts",
+      "packages/dsh-provider-usage/test/unit/common/unit-errsurf.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend-ledger.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend-view.test.ts",
+      "packages/dsh-provider-usage/test/unit/report/unit-report.test.ts",
+      "packages/dsh-provider-usage/test/unit/report/unit-report-executor.test.ts",
+      "packages/dsh-provider-usage/test/unit/routes/unit-routes.test.ts"
     ]
   },
   "jsonReporter": {
