+++ stryker.conf.d/dsh-provider-usage-pipeline.json
@@ -1,15 +1,15 @@
 {
   "$schema": "../node_modules/@stryker-mutator/core/schema/stryker-schema.json",
   "mutate": [
-    "packages/dsh-provider-usage/src/core/history.ts",
-    "packages/dsh-provider-usage/src/pipeline/v2.ts",
-    "packages/dsh-provider-usage/src/hotreload.ts",
+    "packages/dsh-provider-usage/src/domain1/history/history.ts",
+    "packages/dsh-provider-usage/src/domain1/pipeline/v2.ts",
+    "packages/dsh-provider-usage/src/domain1/registry/hotreload.ts",
     "shared/placement-math.js",
-    "packages/dsh-provider-usage/src/path-resolve.ts",
-    "packages/dsh-provider-usage/src/client-logic.ts",
-    "packages/dsh-provider-usage/src/provider-config.ts",
-    "packages/dsh-provider-usage/src/charts.ts",
-    "packages/dsh-provider-usage/src/stats-service.ts",
+    "packages/dsh-provider-usage/src/domain1/registry/path-resolve.ts",
+    "packages/dsh-provider-usage/src/shared/client-logic.ts",
+    "packages/dsh-provider-usage/src/domain1/registry/provider-config.ts",
+    "packages/dsh-provider-usage/src/shared/charts.ts",
+    "packages/dsh-provider-usage/src/domain1/pipeline/stats-service.ts",
     "!packages/dsh-provider-usage/src/client/**",
     "!packages/dsh-provider-usage/src/types.ts"
   ],
@@ -44,20 +44,20 @@
       "./scripts/test/mutation-lib-to-src-hook.mjs"
     ],
     "testFiles": [
-      "packages/dsh-provider-usage/test/unit-apply.test.ts",
-      "packages/dsh-provider-usage/test/unit-config.test.ts",
-      "packages/dsh-provider-usage/test/unit-contract.test.ts",
-      "packages/dsh-provider-usage/test/unit-deepseek-official.test.ts",
-      "packages/dsh-provider-usage/test/unit-errsurf.test.ts",
-      "packages/dsh-provider-usage/test/unit-history.test.ts",
-      "packages/dsh-provider-usage/test/unit-signal-lock.test.ts",
-      "packages/dsh-provider-usage/test/unit-v1.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend-ledger.test.ts",
-      "packages/dsh-provider-usage/test/unit-trend-view.test.ts",
-      "packages/dsh-provider-usage/test/unit-report.test.ts",
-      "packages/dsh-provider-usage/test/unit-routes.test.ts",
-      "packages/dsh-provider-usage/test/unit-report-executor.test.ts"
+      "packages/dsh-provider-usage/test/unit/apply/unit-apply.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-config.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-contract.test.ts",
+      "packages/dsh-provider-usage/test/unit/shared/unit-v1.test.ts",
+      "packages/dsh-provider-usage/test/unit/adapters/unit-deepseek-official.test.ts",
+      "packages/dsh-provider-usage/test/unit/history/unit-history.test.ts",
+      "packages/dsh-provider-usage/test/unit/pipeline/unit-signal-lock.test.ts",
+      "packages/dsh-provider-usage/test/unit/common/unit-errsurf.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend-ledger.test.ts",
+      "packages/dsh-provider-usage/test/unit/trend/unit-trend-view.test.ts",
+      "packages/dsh-provider-usage/test/unit/report/unit-report.test.ts",
+      "packages/dsh-provider-usage/test/unit/report/unit-report-executor.test.ts",
+      "packages/dsh-provider-usage/test/unit/routes/unit-routes.test.ts"
     ]
   },
   "jsonReporter": {
